pushing before moving
