+++ styles.css
@@ -31,9 +31,6 @@ body {
   width: 40px;
 }
 
-/*#ratingpage {
-  background-image: url('images/gear.jpg');
-  background-attachment: fixed;
-  background-position: center;
-  background-size: cover;
-}*/
+#ratingpage {
+  background-color: #37474f;
+}
